requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Linteo y formateo
ruff==0.12.12

# ORM + DB
sqlmodel==0.0.24
psycopg[binary]==3.2.*
alembic==1.13.2

# Validaciones extra
pydantic==2.11.7
python-dotenv==1.0.1

# Seguridad / JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.23.8